Refactor review creation process to delegate file handling to the service and update user ID assignment logic. Modify CreateReviewDto to ensure user ID is required. Enhance file upload service to accept an optional identifier for better file naming.
