+++ src/reviews/reviews.controller.ts
@@ -25,15 +25,11 @@ import {
   ApiBody,
 } from "@nestjs/swagger";
 import { AnyFilesInterceptor } from "@nest-lab/fastify-multer";
-import { FileUploadService } from "../utils/file-upload/fileUpload.service";
 
 @ApiTags("reviews")
 @Controller("reviews")
 export class ReviewsController {
-  constructor(
-    private readonly reviewsService: ReviewsService,
-    private readonly fileUploadService: FileUploadService
-  ) {}
+  constructor(private readonly reviewsService: ReviewsService) {}
 
   @Post()
   @ApiConsumes("multipart/form-data")
@@ -74,66 +70,8 @@ export class ReviewsController {
     @UploadedFiles() files: Array<Express.Multer.File>,
     @Req() req: any
   ) {
-    // Extract files by fieldname (if provided)
-    const profilePic = files?.find((f) => f.fieldname === "profile_picture");
-    const degreeCert = files?.find((f) => f.fieldname === "degree_certificate");
-    const markSheet = files?.find((f) => f.fieldname === "mark_sheet");
-    const studentId = files?.find((f) => f.fieldname === "student_id");
-    const collegeImages = files?.filter(
-      (f) => f.fieldname === "college_images"
-    );
-
-    // Upload each file to S3 (if present) and set URL fields on DTO
-    if (profilePic) {
-      // @ts-ignore - multer File shape from fastify-multer
-      const url = await this.fileUploadService.uploadFile(
-        profilePic as any,
-        "reviews/profile_pictures"
-      );
-      createReviewDto.profile_picture_url = url;
-    }
-
-    if (degreeCert) {
-      const url = await this.fileUploadService.uploadFile(
-        degreeCert as any,
-        "reviews/documents"
-      );
-      createReviewDto.degree_certificate_url = url;
-    }
-
-    if (markSheet) {
-      const url = await this.fileUploadService.uploadFile(
-        markSheet as any,
-        "reviews/documents"
-      );
-      createReviewDto.mark_sheet_url = url;
-    }
-
-    if (studentId) {
-      const url = await this.fileUploadService.uploadFile(
-        studentId as any,
-        "reviews/student-ids"
-      );
-      createReviewDto.student_id_url = url;
-    }
-
-    if (collegeImages && collegeImages.length) {
-      const urls: string[] = [];
-      for (const img of collegeImages) {
-        const url = await this.fileUploadService.uploadFile(
-          img as any,
-          "reviews/college_images"
-        );
-        urls.push(url);
-      }
-      createReviewDto.college_images_urls = urls;
-    }
-
-    if (req?.user?.userId && !createReviewDto.user_id) {
-      createReviewDto.user_id = req.user.userId;
-    }
-
-    return this.reviewsService.create(createReviewDto);
+    // Delegate file handling and creation to service
+    return this.reviewsService.create(createReviewDto, files, req);
   }
 
   @Get()
